udev: immediately fail userspace firmware loading requests
The kernel's userspace firmware fallback mechanicsm has a 60 second
timeout. ath11k appears to be stuck for 60 seconds as it waits for the
userspace helper to fetch "firmware-2.bin". This firmware is not used
on ath11k_ahb devices, so it is normal to be missing. Use a udev rule
to fail the userspace loading, and avoid the 60 second timeout.